refactor(history): extract chart and list components
- Add StoolChartView, LabtestChartView, RecordsList components
- Refactor index.tsx to use extracted components
- Reduce main file from 740 to ~450 lines
